ActionDelegateService
- Added GreetAction with a method to tirgger the action
This method does not have additional operations. It just displays a message.

- Added EventRegistrationAction, also with a method to trigger the action
Here, we can have additonal operations after event has been successfully registered. For now we are just notifiying the user about the events they have registered for, in the console.

FuncDelegateService

Added basic mathematical operations to show how we can use the Func delagate for mathematical calculations.

Usage of these service methods is in the console application.